+++ src/expander.ml
@@ -260,7 +260,8 @@ let expand_and_record acc ~map_exe ~dep_kind ~scope
   in
   let open Build.O in
   match (expansion : Pform.Expansion.t) with
-  | Var (Project_root | First_dep | Deps | Targets | Named_local | Values _)
+  | Var (Project_root | First_dep | Deps | Targets | Target | Named_local
+        | Values _)
   | Macro ((Ocaml_config | Env ), _) -> assert false
   | Var Cc -> add_ddep (cc ~dir).c
   | Var Cxx -> add_ddep (cc ~dir).cxx
@@ -358,6 +359,42 @@ let expand_and_record_deps acc ~(dir : Path.Build.t) ~read_package ~dep_kind
       ~targets_written_by_user ~map_exe ~expand_var ~cc
       t pform syntax_version =
   let res =
+    let module Multiplicity = struct
+      type t = Multiple | One
+    end
+    in
+    let targets ~(multiplicity : Multiplicity.t) =
+      let loc = String_with_vars.Var.loc pform in
+      begin match (targets_written_by_user : Targets.t) with
+      | Infer ->
+        User_error.raise ~loc
+          [ Pp.textf "You cannot use %s with inferred rules."
+              (String_with_vars.Var.describe pform)
+          ]
+      | Forbidden context ->
+        User_error.raise ~loc
+          [ Pp.textf
+              "You cannot use %s in %s."
+              (String_with_vars.Var.describe pform) context
+          ]
+      | Static l ->
+        let value = Value.L.dirs l (* XXX hack to signal no dep *) in
+        match multiplicity, l with
+        | Multiple, _
+        | One, [ _ ]
+          -> Some value
+        | One, [] ->
+          User_error.raise ~loc
+            [ Pp.textf "There is no target." ]
+        | One, _ :: _ :: _ ->
+          User_error.raise ~loc
+            [ Pp.textf
+                "There is more than one target. %s requires there to \
+                 be one unambiguous target."
+                (String_with_vars.Var.describe pform) ]
+
+      end
+    in
     expand_var t pform syntax_version
     |> Option.bind ~f:(function
       | Ok s -> Some s
@@ -368,21 +405,9 @@ let expand_and_record_deps acc ~(dir : Path.Build.t) ~read_package ~dep_kind
           assert false (* these have been expanded statically *)
         | Var (First_dep | Deps | Named_local) -> None
         | Var Targets ->
-          let loc = String_with_vars.Var.loc pform in
-          begin match (targets_written_by_user : Targets.t) with
-          | Infer ->
-            User_error.raise ~loc
-              [ Pp.textf "You cannot use %s with inferred rules."
-                  (String_with_vars.Var.describe pform)
-              ]
-          | Forbidden context ->
-            User_error.raise ~loc
-              [ Pp.textf "You cannot use %s in %s."
-                  (String_with_vars.Var.describe pform) context
-              ]
-          | Static l ->
-            Some (Value.L.dirs l) (* XXX hack to signal no dep *)
-          end
+          targets ~multiplicity:Multiple
+        | Var Target ->
+          targets ~multiplicity:One
         | _ ->
           expand_and_record acc ~map_exe ~dep_kind ~scope:t.scope
             ~expansion_kind:(Dynamic { read_package }) ~dir ~pform